pyproject.toml
[project]
name = "smart-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "openai",
    "pillow",
    "pytest",
    "pytest-json-report",
    "python-dotenv",
    "ruff",
]
